grsec: Apply grsecurity-3.0-3.2.54-201401272346.patch
commit aaf7afa0bc1a4d10383c863dbbeab107168c8228
Merge: df449a2 944ffc0
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Jan 27 23:30:39 2014 -0500

    Merge branch 'pax-stable2' into grsec-stable2

commit 944ffc0c46fb368461f52f3520f7876b8db6e175
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Jan 27 23:30:12 2014 -0500

    backport gcc assert fix from 3.13 constify plugin

 tools/gcc/constify_plugin.c |    4 ++--
 1 files changed, 2 insertions(+), 2 deletions(-)

commit df449a264e347238c119b0fdcdb3e5d83c08d61e
Merge: 1b28d35 99f43f7
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Jan 27 23:16:35 2014 -0500

    Merge branch 'pax-stable2' into grsec-stable2

commit 99f43f743906732e811f76041355a4e460571223
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Jan 27 23:15:52 2014 -0500

    Update to pax-linux-3.2.54-test134.patch:
    - gcc plugin updates
      - ported them to gcc trunk (future 4.9)
      - introduced gcc-common.h to simplify gcc version dependencies
      - updated size overflow hash table from spender
      - fixed kallocstat to detect constant size arguments early
      - fixed constify to preserve type qualifiers on pointer-to-self fields
    - added a few sparse/checker annotations and changes to satisfy gcc's address space logic
    - fixed the TRACE_IRQFLAGS problem reported by minipli again
    - fixed a constify plugin regression, reported by spender
    - updated gcc-common.h
    - fixed the size overflow plugin for gcc 4.9, by Emese Revfy

 arch/x86/ia32/ia32_signal.c               |    6 +-
 arch/x86/kernel/entry_64.S                |   86 ++-----
 arch/x86/kernel/tboot.c                   |    2 +-
 block/compat_ioctl.c                      |    2 +-
 drivers/gpu/drm/drm_crtc.c                |    2 +-
 drivers/media/video/v4l2-compat-ioctl32.c |   10 +-
 drivers/media/video/v4l2-ctrls.c          |    4 +-
 drivers/media/video/v4l2-ioctl.c          |    2 +-
 drivers/mmc/card/block.c                  |    2 +-
 drivers/net/macvtap.c                     |    2 +-
 drivers/video/fbmem.c                     |    2 +-
 fs/compat_ioctl.c                         |    2 +-
 fs/exec.c                                 |    2 +-
 include/linux/videodev2.h                 |    2 +-
 net/9p/client.c                           |    6 +-
 net/socket.c                              |    6 +-
 tools/gcc/Makefile                        |    2 +-
 tools/gcc/checker_plugin.c                |   30 +--
 tools/gcc/colorize_plugin.c               |   62 +++--
 tools/gcc/constify_plugin.c               |  105 ++++----
 tools/gcc/gcc-common.h                    |  267 +++++++++++++++++++
 tools/gcc/kallocstat_plugin.c             |  164 +++++++------
 tools/gcc/kernexec_plugin.c               |  333 ++++++++++++++-----------
 tools/gcc/latent_entropy_plugin.c         |  146 +++++------
 tools/gcc/size_overflow_plugin.c          |  400 +++++++++++++----------------
 tools/gcc/stackleak_plugin.c              |  236 +++++++++--------
 tools/gcc/structleak_plugin.c             |   90 +++----
 27 files changed, 1108 insertions(+), 865 deletions(-)

commit 1b28d3560eaa3295b735bcc4c6d61b320a1cba9a
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Jan 27 20:30:10 2014 -0500

    Make all grsecurity code GPLv2 only for future releases.  Not really
    important as grsecurity is a derivative work of the Linux kernel and
    thus forced to be GPLv2, the "or higher" was superfluous.

 grsecurity/Makefile |    2 +-
 1 files changed, 1 insertions(+), 1 deletions(-)

Signature-tree: 0e1ebe21bb60cb5ae47c5d1147566cd8bdb28072

+++ arch/x86/kernel/tboot.c
@@ -241,7 +241,7 @@ void tboot_shutdown(u32 shutdown_type)
 
 	switch_to_tboot_pt();
 
-	shutdown = (void *)tboot->shutdown_entry;
+	shutdown = (void *)(unsigned long)tboot->shutdown_entry;
 	shutdown();
 
 	/* should not reach here */
